assignment_1/requirements.txt
Django==3.0.3
django-bootstrap3==14.1.0
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.1